requirements.txt
flask
gunicorn
psycopg2-binary    # falls du Postgres benutzt
python-dotenv      # optional, falls du lokal .env nutzt
pytz
Flask-SQLAlchemy
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
werkzeug
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
google-api-python-client
google-auth
google-auth-oauthlib
requests
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
